requirements.txt
anyio==3.6.2
APScheduler==3.10.0
asttokens==2.2.1
certifi==2022.12.7
cheap-repr==0.5.1
click==8.1.3
colorama==0.4.6
executing==1.2.0
fastapi==0.89.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
loguru==0.6.0
msgpack==1.0.4
multidict==6.0.4
nonebot-adapter-onebot==2.2.1
nonebot-plugin-apscheduler==0.2.0
nonebot-plugin-guild-patch==0.2.2
nonebot2==2.0.0rc3
Pillow==9.4.0
pydantic==1.10.4
Pygments==2.14.0
pygtrie==2.5.0
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rfc3986==1.5.0
sentry-sdk==1.14.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
win32-setctime==1.1.0
yarl==1.8.2
